Move tilejson util to preven atlas package import
The shared import moved azure maps into a common module, increasing the
pacakage size significantly.
